ci(semantic-release): add build scripts
Previously, there were no build scripts in the package.json file. These are needed to have a clear
location where files should be copied from for the release, so they have been added. The scripts are
suffixed with :linux with the plan to support win32 versions of the commands at some point.